Update IDispatcher interface:
Clarify the observer role: observe the published actions without any way to interfere in the dispatch process.
Add a IActionMiddleware that will be able to modify the way a selected action is going to be dispatched. For example an action middleware will be able to setup a “Throttling” on some kind of actions.